Skip the process of copying LoDTensorArray in loop_vars in while_loop
test=develop
